+++ datetimepicker/src/androidUnitTest/kotlin/com/kez/picker/PickerFormatRobolectricTest.kt
@@ -23,8 +23,13 @@ import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalTime
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.annotation.Config
 
-class PickerFormatAndroidTest {
+@RunWith(RobolectricTestRunner::class)
+@Config(sdk = [35])
+class PickerFormatRobolectricTest {
 
     @get:Rule
     val composeRule = createComposeRule()